+++ tests/Integration/Execution/MutationExecutor/MorphToManyTest.php
@@ -9,7 +9,7 @@ class MorphToManyTest extends DBTestCase
 {
     protected $schema = '
     type Mutation {
-        createTask(input: CreateTaskInput!): Task @create(flatten: true)
+        createTask(input: CreateTaskInput! @spread): Task @create
     }
     
     input CreateTaskInput {